gcf/python/telemetry-to-firestore/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore >= 0.29.0